Refactoring: functions to parse values from strings
